+++ templates/Models/Group.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\DB;
+use Hamatoma\Laraknife\ViewHelpers;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+
+class Group extends Model
+{
+    use HasFactory;
+    protected $table = 'groups';
+    protected $fillable = [
+        'name',
+        'info',
+        'members'
+    ];
+}